+++ CricketScoreSheetPro.Core/Services/Implementations/TournamentService.cs
@@ -2,6 +2,7 @@
 using CricketScoreSheetPro.Core.Models;
 using CricketScoreSheetPro.Core.Repositories.Interfaces;
 using CricketScoreSheetPro.Core.Services.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -14,12 +15,13 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
 
         public TournamentService(IRepository<Tournament> tournamentRepository, IRepository<UserTournament> usertournamentRepository)
         {
-            _tournamentRepository = tournamentRepository;
-            _usertournamentRepository = usertournamentRepository;
+            _tournamentRepository = tournamentRepository ?? throw new ArgumentNullException($"TournamentRepository is null");
+            _usertournamentRepository = usertournamentRepository ?? throw new ArgumentNullException($"UserTournamentRepository is null");
         }
 
         public async Task<Tournament> AddTournamentAsync(Tournament newTournament)
         {
+            if (newTournament == null) throw new ArgumentNullException($"Tournament is null");
             var tournamentAdd = await _tournamentRepository.CreateAsync(newTournament);
             var newusertournament = new UserTournament
             {
@@ -33,6 +35,8 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
 
         public async Task UpdateTournamentAsync(string tournamentId, Tournament updateTournament)
         {
+            if (updateTournament == null) throw new ArgumentNullException($"Tournament is null");
+            if (string.IsNullOrEmpty(tournamentId)) throw new ArgumentException($"Tournament ID is null");
             await _tournamentRepository.CreateWithIdAsync(tournamentId, updateTournament);
             var updateusertournament = new UserTournament
             {
@@ -45,14 +49,22 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
 
         public async Task<Tournament> GetTournamentAsync(string tournamentId)
         {
+            if (string.IsNullOrEmpty(tournamentId)) throw new ArgumentException($"Tournament ID is null");
             var tournament = await _tournamentRepository.GetItemAsync(tournamentId);
             return tournament;
         }
 
-        public async Task<IList<UserTournament>> GetTournamentsAsync()
+        public async Task<IList<UserTournament>> GetUserTournamentsAsync()
         {
             var usertournaments = await _usertournamentRepository.GetListAsync();
             return usertournaments;
+        }
+
+        public async Task<IList<UserTournament>> GetImportTournamentsAsync(IRepository<UserTournament> importusertournamentRepository)
+        {
+            if (importusertournamentRepository == null) throw new ArgumentNullException($"Import UserTournamentRepo is null");
+            var importusertournaments = await importusertournamentRepository.GetListAsync();
+            return importusertournaments;
         }
 
         public async Task DeleteAllTournamentsAsync()
@@ -63,6 +75,7 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
 
         public async Task DeleteTournamentAsync(string tournamentId)
         {
+            if (string.IsNullOrEmpty(tournamentId)) throw new ArgumentException($"Tournament ID is null");
             await _tournamentRepository.DeleteByIdAsync(tournamentId);
             await _usertournamentRepository.DeleteByIdAsync(tournamentId);
         }
